Delete modal dialog ile silme onaylama

+++ ECommerceClient/src/app/directives/admin/delete.ts
@@ -1,8 +1,10 @@
-import { Directive, ElementRef, EventEmitter, HostListener, Input, Output, Renderer2 } from '@angular/core';
+import {Directive, ElementRef, EventEmitter, HostListener, inject, Input, Output, Renderer2} from '@angular/core';
 import { HttpClientService } from '../../services/common/http-client';
 import { ProductService } from '../../services/common/models/product';
 import { Base, SpinnerType } from '../../base/base';
 import { NgxSpinnerService } from 'ngx-spinner';
+import {MatDialog} from '@angular/material/dialog';
+import {DeleteDialog} from '../../dialogs/delete-dialog/delete-dialog';
 
 declare var $: any;
 
@@ -11,9 +13,10 @@ declare var $: any;
 })
 export class DeleteDirective {
 
+  readonly dialog = inject(MatDialog);
   constructor(
-    private element: ElementRef, 
-    private _renderer: Renderer2, 
+    private element: ElementRef,
+    private _renderer: Renderer2,
     private productService: ProductService ,
     private spinner: NgxSpinnerService
   ) {
@@ -28,12 +31,27 @@ export class DeleteDirective {
    @Output() hasanCallBack: EventEmitter<any> = new EventEmitter();
 
    @HostListener("click") // ilgili direciv'in(appDelete) kullanıldığı dom nesnesine tıklanıldığında hangi olay verildiyse o olay gerçekleşir
-   async onclick() {
-    this.spinner.show(SpinnerType.BALL_SPIN_CLOCKWİSE_FADE_ROTATING);
-    const td: HTMLTableCellElement = this.element.nativeElement;
-    await this.productService.delete(this.id)
-    $(td.parentElement).fadeOut(1500, () => {
-      this.hasanCallBack.emit();
-    });
+
+   onclick() {
+    this.openDialog('300ms', '300ms');
    }
+
+  openDialog(enterAnimationDuration: string, exitAnimationDuration: string): void {
+    const dialogRef = this.dialog.open(DeleteDialog, {
+      width: '250px',
+      enterAnimationDuration,
+      exitAnimationDuration,
+    });
+
+    dialogRef.afterClosed().subscribe(async result => {
+      if (result) {
+        this.spinner.show(SpinnerType.BALL_SPIN_CLOCKWİSE_FADE_ROTATING);
+        const td: HTMLTableCellElement = this.element.nativeElement;
+        await this.productService.delete(this.id)
+        $(td.parentElement).fadeOut(1000, () => {
+          this.hasanCallBack.emit();
+        });
+      }
+    });
+  }
 }
